# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AnimeThemes.moe Api Documentation",
        "description": "AnimeThemes.moe Laravel RESTful API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/year/{year}": {
            "get": {
                "tags": [
                    "Anime"
                ],
                "summary": "Get listing of Anime of year by season",
                "description": "Returns listing of Anime of year by season",
                "operationId": "getAnimeOfYearBySeason",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Comma-separated list of included related resources. Allowed list is synonyms, series, themes, themes.entries, themes.entries.videos, themes.song, themes.song.artists & externalResources.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "synonyms,series"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "summer": {
                                            "description": "Display a listing of anime of year by season.",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AnimeResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AnimeResource": {
                "title": "Anime",
                "description": "Anime Resource",
                "properties": {
                    "id": {
                        "description": "Primary Key",
                        "type": "integer",
                        "example": 197
                    },
                    "name": {
                        "description": "The Primary Title of the Anime",
                        "type": "string",
                        "example": "Bakemonogatari"
                    },
                    "alias": {
                        "description": "URL Slug & Model Route Key",
                        "type": "string",
                        "example": "bakemonogatari"
                    },
                    "year": {
                        "description": "The Year in which the Anime Premiered",
                        "type": "integer",
                        "example": 2009
                    },
                    "season": {
                        "description": "The Season in which the Anime Premiered",
                        "type": "string",
                        "enum": [
                            "Winter",
                            "Spring",
                            "Summer",
                            "Fall"
                        ],
                        "example": "Summer"
                    },
                    "created_at": {
                        "description": "The Resource Creation Timestamp",
                        "type": "string",
                        "example": "2020-08-15T05:30:43.000000Z"
                    },
                    "updated_at": {
                        "description": "The Resource Last Updated Timestamp",
                        "type": "string",
                        "example": "2020-08-15T05:37:25.000000Z"
                    },
                    "synonyms": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "Primary Key",
                                    "type": "integer",
                                    "example": 1464
                                },
                                "text": {
                                    "description": "For alternative titles, licensed titles, common abbreviations and/or shortenings",
                                    "type": "string",
                                    "example": "Monstory"
                                },
                                "created_at": {
                                    "description": "The Resource Creation Timestamp",
                                    "type": "string",
                                    "example": "2020-08-15T05:43:02.000000Z"
                                },
                                "updated_at": {
                                    "description": "The Resource Last Updated Timestamp",
                                    "type": "string",
                                    "example": "2020-08-15T05:43:02.000000Z"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "themes": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "Primary Key",
                                    "type": "integer",
                                    "example": 3102
                                },
                                "type": {
                                    "description": "Is this an OP or an ED?",
                                    "type": "string",
                                    "enum": [
                                        "OP",
                                        "ED"
                                    ],
                                    "example": "OP"
                                },
                                "sequence": {
                                    "description": "Numeric ordering of theme",
                                    "type": "integer",
                                    "example": "1"
                                },
                                "group": {
                                    "description": "For separating sequences belonging to dubs, rebroadcasts, remasters, etc",
                                    "type": "string",
                                    "example": ""
                                },
                                "slug": {
                                    "description": "URL Slug",
                                    "type": "bool",
                                    "example": "OP1"
                                },
                                "created_at": {
                                    "description": "The Resource Creation Timestamp",
                                    "type": "string",
                                    "example": "2020-08-15T05:43:02.000000Z"
                                },
                                "updated_at": {
                                    "description": "The Resource Last Updated Timestamp",
                                    "type": "string",
                                    "example": "2020-08-15T05:43:02.000000Z"
                                },
                                "song": {
                                    "properties": {
                                        "id": {
                                            "description": "Primary Key",
                                            "type": "integer",
                                            "example": 3102
                                        },
                                        "title": {
                                            "description": "The title of the song",
                                            "type": "string",
                                            "example": "staple stable"
                                        },
                                        "created_at": {
                                            "description": "The Resource Creation Timestamp",
                                            "type": "string",
                                            "example": "2020-08-15T05:43:02.000000Z"
                                        },
                                        "updated_at": {
                                            "description": "The Resource Last Updated Timestamp",
                                            "type": "string",
                                            "example": "2020-08-15T05:43:02.000000Z"
                                        },
                                        "artists": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Primary Key",
                                                        "type": "integer",
                                                        "example": 53
                                                    },
                                                    "name": {
                                                        "description": "The Primary Name of the Artist",
                                                        "type": "string",
                                                        "example": "Chiwa Saito"
                                                    },
                                                    "alias": {
                                                        "description": "URL Slug & Model Route Key",
                                                        "type": "string",
                                                        "example": "chiwa_saito"
                                                    },
                                                    "as": {
                                                        "description": "Used in place of the Artist name if the performance is made as a character or group/unit member",
                                                        "type": "string",
                                                        "example": "Hitagi Senjougahara"
                                                    },
                                                    "created_at": {
                                                        "description": "The Resource Creation Timestamp",
                                                        "type": "string",
                                                        "example": "2020-08-15T05:55:55.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "description": "The Resource Last Updated Timestamp",
                                                        "type": "string",
                                                        "example": "2020-08-15T05:55:55.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                },
                                "entries": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "Primary Key",
                                                "type": "integer",
                                                "example": 3518
                                            },
                                            "version": {
                                                "description": "The Version number of the Theme",
                                                "type": "integer",
                                                "example": ""
                                            },
                                            "episodes": {
                                                "description": "The range(s) of episodes that the theme entry is used",
                                                "type": "string",
                                                "example": "1-2, 12"
                                            },
                                            "nsfw": {
                                                "description": "Does the entry include Not Safe For Work content?",
                                                "type": "bool",
                                                "example": "false"
                                            },
                                            "spoiler": {
                                                "description": "Does the entry include content that spoils the show?",
                                                "type": "bool",
                                                "example": "false"
                                            },
                                            "notes": {
                                                "description": "Any additional information not included in other fields that may be useful",
                                                "type": "string",
                                                "example": ""
                                            },
                                            "created_at": {
                                                "description": "The Resource Creation Timestamp",
                                                "type": "string",
                                                "example": "2020-08-15T05:43:02.000000Z"
                                            },
                                            "updated_at": {
                                                "description": "The Resource Last Updated Timestamp",
                                                "type": "string",
                                                "example": "2020-08-15T05:43:02.000000Z"
                                            },
                                            "videos": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "description": "Primary Key",
                                                            "type": "integer",
                                                            "example": 2615
                                                        },
                                                        "basename": {
                                                            "description": "The basename of the Video",
                                                            "type": "string",
                                                            "example": "Bakemonogatari-OP1.webm"
                                                        },
                                                        "filename": {
                                                            "description": "The filename of the Video",
                                                            "type": "string",
                                                            "example": "Bakemonogatari-OP1"
                                                        },
                                                        "path": {
                                                            "description": "The path of the Video in storage",
                                                            "type": "string",
                                                            "example": "2009/Summer/Bakemonogatari-OP1.webm"
                                                        },
                                                        "resolution": {
                                                            "description": "Frame height of the video",
                                                            "type": "integer",
                                                            "example": 1080
                                                        },
                                                        "nc": {
                                                            "description": "Is the video creditless?",
                                                            "type": "bool",
                                                            "example": "true"
                                                        },
                                                        "subbed": {
                                                            "description": "Does the video include subtitles of dialogue?",
                                                            "type": "bool",
                                                            "example": "false"
                                                        },
                                                        "lyrics": {
                                                            "description": "Does the video include subtitles for song lyrics?",
                                                            "type": "bool",
                                                            "example": "false"
                                                        },
                                                        "uncen": {
                                                            "description": "Is the video an uncensored version of a censored sequence?",
                                                            "type": "bool",
                                                            "example": "false"
                                                        },
                                                        "source": {
                                                            "description": "Where did this video come from?",
                                                            "type": "string",
                                                            "enum": [
                                                                "WEB",
                                                                "RAW",
                                                                "BD",
                                                                "DVD",
                                                                "VHS"
                                                            ],
                                                            "example": "BD"
                                                        },
                                                        "overlap": {
                                                            "description": "The degree to which the sequence and episode content overlap",
                                                            "type": "string",
                                                            "enum": [
                                                                "None",
                                                                "Transition",
                                                                "Over"
                                                            ],
                                                            "example": "None"
                                                        },
                                                        "created_at": {
                                                            "description": "The Resource Creation Timestamp",
                                                            "type": "string",
                                                            "example": "2020-08-15T05:30:43.000000Z"
                                                        },
                                                        "updated_at": {
                                                            "description": "The Resource Last Updated Timestamp",
                                                            "type": "string",
                                                            "example": "2020-08-15T05:30:43.000000Z"
                                                        },
                                                        "link": {
                                                            "description": "The link to the video stream",
                                                            "type": "string",
                                                            "example": "https://animethemes.moe/video/Bakemonogatari-OP1.webm"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    },
                    "series": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "Primary Key",
                                    "type": "integer",
                                    "example": 114
                                },
                                "name": {
                                    "description": "The Primary Name of the Series",
                                    "type": "string",
                                    "example": "Monogatari"
                                },
                                "alias": {
                                    "description": "URL Slug & Model Route Key",
                                    "type": "string",
                                    "example": "monogatari"
                                },
                                "created_at": {
                                    "description": "The Resource Creation Timestamp",
                                    "type": "string",
                                    "example": "2020-08-15T06:53:20.000000Z"
                                },
                                "updated_at": {
                                    "description": "The Resource Last Updated Timestamp",
                                    "type": "string",
                                    "example": "2020-08-15T06:53:20.000000Z"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "resources": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "Primary Key",
                                    "type": "integer",
                                    "example": 1018
                                },
                                "link": {
                                    "description": "The URL of the resource",
                                    "type": "string",
                                    "example": "https://myanimelist.net/anime/5081/"
                                },
                                "external_id": {
                                    "description": "The identifier used by the external site",
                                    "type": "integer",
                                    "example": "5081"
                                },
                                "type": {
                                    "description": "The site that we are linking to",
                                    "type": "string",
                                    "enum": [
                                        "Official Website",
                                        "Twitter",
                                        "aniDB",
                                        "AniList",
                                        "Anime-Planet",
                                        "Anime News Network",
                                        "Kitsu",
                                        "MyAnimeList",
                                        "Wiki"
                                    ],
                                    "example": "MyAnimeList"
                                },
                                "as": {
                                    "description": "Used to distinguish resources that map to the same artist or anime",
                                    "type": "string",
                                    "example": ""
                                },
                                "created_at": {
                                    "description": "The Resource Creation Timestamp",
                                    "type": "string",
                                    "example": "2020-08-15T05:30:43.000000Z"
                                },
                                "updated_at": {
                                    "description": "The Resource Last Updated Timestamp",
                                    "type": "string",
                                    "example": "2020-08-15T05:37:25.000000Z"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
    "security": [
        []
    ]
}